Move to int8
If we want to set a delay of ~30 days - this will cause an issue because
the functions take int4s.

+++ src/command/message-create.ts
@@ -82,8 +82,8 @@ export class MessageCreateCommand {
                 $2, 
                 $3, 
                 $4,
-                $5::INTEGER, 
-                $6::INTEGER
+                $5::BIGINT, 
+                $6::BIGINT
             )
         `.value, [
             this.id,
